% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{graph}
\alias{graph}
\title{Plots the ggplot file}
\usage{
graph(data, column, column_1)
}
\arguments{
\item{data}{selects the data column}

\item{column}{selects the first column from which the data is taken and plotted on scatter plot}

\item{column_1}{slectes the second column which is used to plot the scatter plot}
}
\description{
Using 3 inputs, this function to go through data and plots the selected columns as a scatter plot.
}
