% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmplot.R
\name{lm_function}
\alias{lm_function}
\title{This function does linear modelling.
This function describes a relation between dependent and independent columns.}
\usage{
lm_function(df, dependent, independent_cols)
}
\arguments{
\item{df}{Crabs Data set}

\item{dependent}{The dependent values}

\item{independent_cols}{The independent values}
}
\description{
This function does linear modelling.
This function describes a relation between dependent and independent columns.
}
