% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meme.R
\name{meme}
\alias{meme}
\title{This function shows a meme whenever a function doesn't work.}
\usage{
meme(img)
}
\arguments{
\item{img}{Calls the image file.}
}
\description{
This function shows a meme whenever a function doesn't work.
}
