% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutatecolumns.R
\name{mutate_size}
\alias{mutate_size}
\title{Changes the number of columns by adding or combine two columns into one.}
\usage{
mutate_size(data)
}
\arguments{
\item{data}{selects the data column}
}
\description{
This function will take 3 columns from the crabs data and then calculate estimated volume using the 3 different measurements.
}
