% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveNA.R
\name{remove_nas}
\alias{remove_nas}
\title{Removes na in a data}
\usage{
remove_nas(data, column_name)
}
\arguments{
\item{data}{selects the data column}

\item{column}{selects the column from which the data is taken}
}
\description{
Using a single data input, this function to go through each column and cleaning it out by removing NA value.
}
